Conflicts:
	config/java_opts.yml

+++ spec/application_helper.rb
@@ -1,6 +1,6 @@
 # Encoding: utf-8
 # Cloud Foundry Java Buildpack
-# Copyright 2013 the original author or authors.
+# Copyright 2013-2015 the original author or authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -65,8 +65,12 @@ shared_context 'application_helper' do
     application
   end
 
-  after do
-    FileUtils.rm_rf app_dir
+  after do |example|
+    if example.metadata[:no_cleanup]
+      puts "Application Directory: #{app_dir}"
+    else
+      FileUtils.rm_rf app_dir
+    end
   end
 
 end
