+++ spec/integration_helper.rb
@@ -1,6 +1,6 @@
 # Encoding: utf-8
 # Cloud Foundry Java Buildpack
-# Copyright 2013 the original author or authors.
+# Copyright 2013-2015 the original author or authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -38,8 +38,12 @@ shared_context 'integration_helper' do
     FileUtils.cp_r "spec/fixtures/#{buildpack_fixture.chomp}/.", buildpack_dir if buildpack_fixture
   end
 
-  after do
-    FileUtils.rm_rf buildpack_dir
+  after do |example|
+    if example.metadata[:no_cleanup]
+      puts "Buildpack Directory: #{buildpack_dir}"
+    else
+      FileUtils.rm_rf buildpack_dir
+    end
   end
 
   def run(command)
